V1.0

+++ backend/evolvehub-common/src/main/java/org/evolve/common/web/page/PageResponse.java
@@ -16,6 +16,6 @@ import java.util.List;
 public record PageResponse<T>(
         List<T> records,
         long total,
-        int pageNum,
-        int pageSize) {
+        long pageNum,
+        long pageSize) {
 }
